chg: splitted custom written logics in different file
